requirements.txt
simpleflake
requests
libmc
msgpack-python
pyquery
Flask
Flask-SQLAlchemy
Flask-Script
Flask-Migrate
Flask-Testing
